Refactor Gateway connection
