--- src/Gateway/Objects/D/Hello.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Ragnarok\Fenrir\Gateway\Objects\D;
-
-class Hello
-{
-    public int $heartbeat_interval;
-}
